+++ src/intrinsics/simd.rs
@@ -53,7 +53,7 @@ pub(super) fn codegen_simd_intrinsic_call<'tcx>(
 
         // simd_shuffle32<T, U>(x: T, y: T, idx: [u32; 32]) -> U
         _ if intrinsic.starts_with("simd_shuffle"), (c x, c y, o idx) {
-            let n: u32 = intrinsic["simd_shuffle".len()..].parse().unwrap();
+            let n: u16 = intrinsic["simd_shuffle".len()..].parse().unwrap();
 
             assert_eq!(x.layout(), y.layout());
             let layout = x.layout();
@@ -85,8 +85,8 @@ pub(super) fn codegen_simd_intrinsic_call<'tcx>(
                         fx.tcx.data_layout.endian,
                         &idx_bytes[4*i.. 4*i + 4],
                     ).expect("read_target_uint");
-                    u32::try_from(idx).expect("try_from u32")
-                }).collect::<Vec<u32>>()
+                    u16::try_from(idx).expect("try_from u32")
+                }).collect::<Vec<u16>>()
             };
 
             for &idx in &indexes {
